Updates pharmaRTF to be compatible with huxtable v5. Also adds in compatibility for header_rows = 0.
huxtable was added depandancy rather than a suggestion as the package won't function without it.

Version in DESCRIPTION was updated to 0.1.1.

Added a NEWS.md file to explain changes in the new update.

helpers.R was updated due to changes in how huxtable subsets tables.

get_column_headers.huxtable was updated to only print if header_rows aren't 0.

Most tests were updated due to the default in add_colnames being changed to TRUE.